Fix some Julia 0.4 deprecations related to {} and [] for #65
Not sure how to fix the concatenation deprecations, though.
As far as I can tell, the code is not concatenating anything,
just using `[a, b]` syntax to construct vectors from elements.

Not sure how to fix the deprecation for `Dict` syntax either,
as the new syntax is invalid in Julia 0.3.